More review comments
* add comments for immutable reference configs
* reorganize code for better readability
