+++ validation/standard/standard-immutables.toml
@@ -1,3 +1,13 @@
+# Immutable references indicate the positions of the immutables in the bytecode. The combined contract JSON generated by
+# the compiler at the stated tag below will contain such references (See example below). To add new contracts with immutables/
+# update a contracts's immutables, paste the snippet of `"immutableReferences"` from the contract json file. Immutable values
+# are likely different in every deployment of the contract. So we mask the value of the immutable references in the bytecode
+# before computing any hash for checking
+#
+# Example: ethereum-optimism/optimism/packages/contracts-bedrock/forge-artifacts/MIPS.sol/MIPS.json
+#
+# "immutableReferences":{"85798":[{"start":178,"length":32},{"start":1771,"length":32}]}}
+#
 ["op-contracts/v1.4.0"] # Fault Proof contract AST https://github.com/ethereum-optimism/optimism/releases/tag/op-contracts%2Fv1.4.0
 anchor_state_registry = '{"92829":[{"start":387,"length":32},{"start":828,"length":32},{"start":2296,"length":32}]}'
 delayed_weth = '{"97827":[{"start":831,"length":32},{"start":4133,"length":32}]}'
